backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
Flask-CORS==3.0.10
psycopg2-binary==2.9.10
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
alembic==1.7.1
pytest==6.2.5
gunicorn==20.1.0
bcrypt==4.0.1
email-validator==2.0.0.post2
pytesseract==0.3.10
Pillow==10.0.0
SpeechRecognition==3.10.0
jieba==0.42.1
vosk==0.3.45
dashscope
oss2
aliyun-python-sdk-core>=2.13.36
aliyun-python-sdk-nls-cloud-meta>=1.0.0
alibabacloud-ocr-api20210707>=1.0.0
alibabacloud-tea-openapi>=0.3.8 